# src/markweave/http/routers/conversions.py
"""Conversion submission, lifecycle, and result routes."""

from datetime import UTC, datetime
from typing import Annotated
from urllib.parse import quote
from uuid import UUID

from fastapi import (
    APIRouter,
    Depends,
    File,
    Form,
    Header,
    Query,
    Request,
    Response,
    UploadFile,
    status,
)
from starlette.concurrency import run_in_threadpool

from markweave.auth.models import Role, User
from markweave.http.dependencies import HttpDependencies
from markweave.http.errors import error_responses
from markweave.http.responses import conversion_response, template_response
from markweave.http.schemas import (
    ConversionOptionsResponse,
    ConversionPageResponse,
    ConversionResponse,
)
from markweave.jobs.errors import JobRequestError
from markweave.jobs.models import (
    JobOutput,
    JobPage,
    JobRequest,
    source_kind_for_filename,
)
from markweave.observability import CORRELATION_HEADER, CORRELATION_STATE_KEY
from markweave.persistence.errors import PersistenceError
from markweave.presentations.models import PresentationDialect, PresentationOptions
from markweave.templates.models import TemplateKind, TemplateSelectionSource
from markweave.version import VERSION

COMPONENT_VERSIONS = (
    ("chromium", "151.0.7922.173"),
    ("libreoffice", "26.2.5.2"),
    ("md-converter", VERSION),
    ("mermaid-cli", "11.16.0"),
    ("pandoc", "3.10.2"),
)

_RESULT_EXTENSIONS = {
    JobOutput.DOCX: "docx",
    JobOutput.PDF: "pdf",
    JobOutput.BOTH: "zip",
    JobOutput.PPTX: "pptx",
    JobOutput.PPTX_BUNDLE: "zip",
}


def _result_content_disposition(
    source_filename: str | None,
    output: JobOutput,
    *,
    fallback_stem: str,
) -> str:
    """Build a safe attachment header preserving the uploaded filename stem."""

    stem = (
        source_filename.rsplit(".", maxsplit=1)[0]
        if source_filename is not None
        else fallback_stem
    )
    filename = f"{stem}.{_RESULT_EXTENSIONS[output]}"
    encoded_filename = quote(filename, safe="")
    if encoded_filename != filename:
        return f"attachment; filename*=UTF-8''{encoded_filename}"
    return f'attachment; filename="{filename}"'


def build_router(dependencies: HttpDependencies) -> APIRouter:
    """Build conversion routes bound to one application."""

    router = APIRouter()
    settings = dependencies.settings
    components = dependencies.components

    @router.get(
        "/api/v1/conversion-options",
        response_model=ConversionOptionsResponse,
        tags=["conversions"],
        responses=error_responses(401, 503),
    )
    def get_conversion_options(
        response: Response,
        actor: Annotated[User, Depends(dependencies.current_user)],
        template_kind: TemplateKind = TemplateKind.DOCX,
    ) -> ConversionOptionsResponse:
        response.headers["Cache-Control"] = "no-store"
        if template_kind is TemplateKind.PPTX:
            template, source = None, TemplateSelectionSource.PANDOC_DEFAULT
        else:
            template, source = dependencies.template_runtime().resolve_with_source(
                actor
            )
        version_id = template.current_version_id if template is not None else None
        if template is not None and version_id is None:
            raise PersistenceError
        return ConversionOptionsResponse(
            conversion_upload_max_bytes=settings.conversion_upload_max_bytes,
            resolved_template=(
                template_response(template, dependencies.authentication)
                if template is not None
                else None
            ),
            template_version_id=version_id,
            selection_source=source,
        )

    @router.post(
        "/api/v1/conversions",
        response_model=ConversionResponse,
        status_code=status.HTTP_202_ACCEPTED,
        tags=["conversions"],
        responses=error_responses(401, 403, 409, 413, 422, 429, 503),
    )
    async def create_conversion(  # noqa: PLR0913, PLR0917 - FastAPI fields
        request: Request,
        response: Response,
        actor: Annotated[User, Depends(dependencies.mutation_actor)],
        source: Annotated[UploadFile, File()],
        output: Annotated[JobOutput, Form()],
        template_id: Annotated[UUID | None, Form()] = None,
        template_version_id: Annotated[UUID | None, Form()] = None,
        idempotency_key: Annotated[str | None, Header(alias="Idempotency-Key")] = None,
        presentation_dialect: Annotated[PresentationDialect | None, Form()] = None,
        slide_level: Annotated[int | None, Form(ge=1, le=6)] = None,
    ) -> ConversionResponse:
        correlation_id = getattr(request.state, CORRELATION_STATE_KEY)
        if source.filename is None:
            await source.close()
            raise JobRequestError
        source_filename = source.filename
        try:
            source_kind = source_kind_for_filename(source_filename)
        except ValueError:
            await source.close()
            raise JobRequestError from None
        try:
            content = await source.read(settings.conversion_upload_max_bytes + 1)
        finally:
            await source.close()
        if not content or len(content) > settings.conversion_upload_max_bytes:
            raise JobRequestError
        await run_in_threadpool(components.scanner.scan, content)
        try:
            job, _replayed = await run_in_threadpool(
                components.jobs.submit,
                JobRequest(
                    owner_id=actor.id,
                    source=content,
                    template_id=template_id,
                    template_version_id=template_version_id,
                    output=output,
                    component_versions=COMPONENT_VERSIONS,
                    now=datetime.now(UTC),
                    correlation_id=correlation_id,
                    source_filename=source_filename,
                    source_kind=source_kind,
                    presentation_options=(
                        PresentationOptions(
                            presentation_dialect or PresentationDialect.AUTO,
                            slide_level if slide_level is not None else 2,
                        )
                        if output.is_presentation
                        or presentation_dialect is not None
                        or slide_level is not None
                        else None
                    ),
                ),
                idempotency_key,
            )
        except ValueError:
            raise JobRequestError from None
        response.headers["Location"] = f"/api/v1/conversions/{job.id}"
        response.headers[CORRELATION_HEADER] = job.correlation_id
        response.headers["Retry-After"] = str(settings.conversion_retry_after_seconds)
        return conversion_response(job)

    @router.get(
        "/api/v1/conversions",
        response_model=ConversionPageResponse,
        tags=["conversions"],
        responses=error_responses(401, 422, 503),
    )
    def list_conversions(
        actor: Annotated[User, Depends(dependencies.current_user)],
        offset: Annotated[int, Query(ge=0)] = 0,
        limit: Annotated[int, Query(ge=1, le=100)] = 50,
    ) -> ConversionPageResponse:
        page: JobPage = components.jobs.list_owner(actor.id, offset=offset, limit=limit)
        return ConversionPageResponse(
            items=tuple(conversion_response(job) for job in page.items),
            total=page.total,
            offset=page.offset,
            limit=page.limit,
        )

    @router.get(
        "/api/v1/conversions/{job_id}",
        response_model=ConversionResponse,
        tags=["conversions"],
        responses=error_responses(401, 404, 422, 503),
    )
    def get_conversion(
        job_id: UUID,
        actor: Annotated[User, Depends(dependencies.current_user)],
    ) -> ConversionResponse:
        return conversion_response(
            components.jobs.get_visible(
                job_id,
                actor_id=actor.id,
                actor_is_admin=actor.role is Role.ADMIN,
            )
        )

    @router.delete(
        "/api/v1/conversions/{job_id}",
        response_model=ConversionResponse,
        tags=["conversions"],
        responses=error_responses(401, 403, 404, 422, 503),
    )
    def cancel_conversion(
        job_id: UUID,
        actor: Annotated[User, Depends(dependencies.mutation_actor)],
    ) -> ConversionResponse:
        return conversion_response(
            components.jobs.cancel(
                job_id,
                actor_id=actor.id,
                actor_is_admin=actor.role is Role.ADMIN,
                now=datetime.now(UTC),
            )
        )

    @router.get(
        "/api/v1/conversions/{job_id}/result",
        response_class=Response,
        tags=["conversions"],
        responses={
            200: {
                "description": "Immutable conversion result",
                "content": {
                    "application/octet-stream": {
                        "schema": {"type": "string", "format": "binary"}
                    }
                },
            },
            **error_responses(401, 404, 409, 422, 503),
        },
    )
    def download_conversion(
        job_id: UUID,
        actor: Annotated[User, Depends(dependencies.current_user)],
    ) -> Response:
        job, content = components.jobs.download(
            job_id,
            actor_id=actor.id,
            actor_is_admin=actor.role is Role.ADMIN,
        )
        return Response(
            content,
            media_type="application/octet-stream",
            headers={
                "Content-Disposition": _result_content_disposition(
                    job.source_filename,
                    job.output,
                    fallback_stem=f"conversion-{job.id}",
                ),
                "Cache-Control": "private, no-store",
                "X-Content-Type-Options": "nosniff",
            },
        )

    @router.get(
        "/api/v1/conversions/{job_id}/result/manifest",
        response_class=Response,
        tags=["conversions"],
        responses={
            200: {
                "description": "Canonical PDF traceability manifest",
                "content": {"application/json": {"schema": {"type": "object"}}},
            },
            **error_responses(401, 404, 409, 422, 503),
        },
    )
    def download_conversion_manifest(
        job_id: UUID,
        actor: Annotated[User, Depends(dependencies.current_user)],
    ) -> Response:
        job, content = components.jobs.download_manifest(
            job_id,
            actor_id=actor.id,
            actor_is_admin=actor.role is Role.ADMIN,
        )
        return Response(
            content,
            media_type="application/json",
            headers={
                "Content-Disposition": (
                    f'attachment; filename="conversion-{job.id}-traceability.json"'
                ),
                "Cache-Control": "private, no-store",
                "X-Content-Type-Options": "nosniff",
            },
        )

    return router