# tests/fixtures/t41_http_contract/openapi.json
{
  "components": {
    "schemas": {
...
      "ConversionOptionsResponse": {
        "description": "Authoritative limits and immutable template selection for conversion.",
        "properties": {
          "conversion_upload_max_bytes": {
            "exclusiveMinimum": 0.0,
            "title": "Conversion Upload Max Bytes",
            "type": "integer"
          },
          "resolved_template": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TemplateResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "selection_source": {
            "$ref": "#/components/schemas/TemplateSelectionSource"
          },
          "template_version_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Version Id"
          }
        },
        "required": [
          "conversion_upload_max_bytes",
          "resolved_template",
          "template_version_id",
          "selection_source"
        ],
        "title": "ConversionOptionsResponse",
        "type": "object"
      },
      "ConversionPageResponse": {
        "description": "Paginated owner conversion response.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ConversionResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "ConversionPageResponse",
        "type": "object"
      },
...
      "ErrorDetail": {
        "description": "Stable machine-readable functional error detail.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "Stable error envelope used for every expected HTTP failure.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TemplateKind": {
        "description": "Immutable reference document family.",
        "enum": [
          "docx",
          "pptx"
        ],
        "title": "TemplateKind",
        "type": "string"
      },
...
      "TemplateSelectionSource": {
        "description": "Authoritative source of the template selected for a conversion.",
        "enum": [
          "pandoc_default",
          "preferred",
          "system_fallback"
        ],
        "title": "TemplateSelectionSource",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "description": "Opaque authenticated Markweave session cookie.",
        "in": "cookie",
        "name": "md_converter_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Markdown Converter API",
    "version": "0.7.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/conversion-options": {
      "get": {
        "operationId": "get_conversion_options_api_v1_conversion_options_get",
        "parameters": [
          {
            "in": "query",
            "name": "template_kind",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/TemplateKind",
              "default": "docx"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversionOptionsResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication failed or is required",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The service is not ready",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Get Conversion Options",
        "tags": [
          "conversions"
        ]
      }
    },
    "/api/v1/conversions": {
      "get": {
        "operationId": "list_conversions_api_v1_conversions_get",
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversionPageResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication failed or is required",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The request is invalid",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The service is not ready",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "List Conversions",
        "tags": [
          "conversions"
        ]
      },
      "post": {
        "operationId": "create_conversion_api_v1_conversions_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_conversion_api_v1_conversions_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversionResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication failed or is required",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The operation is forbidden",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The request conflicts with current state",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The request body is too large",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The request is invalid",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The caller has exceeded a configured quota",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The service is not ready",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Create Conversion",
        "tags": [
          "conversions"
        ]
      }
    },
    "/api/v1/conversions/{job_id}": {
      "delete": {
        "operationId": "cancel_conversion_api_v1_conversions__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversionResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication failed or is required",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The operation is forbidden",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The requested resource was not found",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The request is invalid",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The service is not ready",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Cancel Conversion",
        "tags": [
          "conversions"
        ]
      },
      "get": {
        "operationId": "get_conversion_api_v1_conversions__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversionResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication failed or is required",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The requested resource was not found",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The request is invalid",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The service is not ready",
            "headers": {
              "X-Correlation-ID": {
                "description": "Server-generated request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Get Conversion",
        "tags": [
          "conversions"
        ]
      }
    },
...
  },
  "security": [
    {
      "SessionCookie": []
    }
  ]
}